+++ example/mqtt_server_client_secure.dart
@@ -63,6 +63,9 @@ Future<int> main() async {
   /// from the broker.
   client.pongCallback = pong;
 
+  /// Set an on bad certificate callback, note that the parameter is needed.
+  client.onBadCertificate = (dynamic a) => true;
+
   /// Create a connection message to use or use the default one. The default one sets the
   /// client identifier, any supplied username/password, the default keepalive interval(60s)
   /// and clean session, an example of a specific one below.
@@ -166,7 +169,8 @@ void onSubscribed(String topic) {
 /// The unsolicited disconnect callback
 void onDisconnected() {
   print('EXAMPLE::OnDisconnected client callback - Client disconnection');
-  if (client.connectionStatus.returnCode == MqttConnectReturnCode.solicited) {
+  if (client.connectionStatus.disconnectionOrigin ==
+      MqttDisconnectionOrigin.solicited) {
     print('EXAMPLE::OnDisconnected callback is solicited, this is correct');
   }
   exit(-1);